move citizen safetConcern mapper outside controller
