+++ src/main/java/uk/gov/hmcts/reform/prl/controllers/citizen/mapper/CaseDataRespondentDetailsElementsMapper.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.prl.mapper.citizen;
+package uk.gov.hmcts.reform.prl.controllers.citizen.mapper;
 
 import uk.gov.hmcts.reform.prl.enums.DontKnow;
 import uk.gov.hmcts.reform.prl.enums.Gender;
