+++ src/main/java/uk/gov/hmcts/reform/prl/rpa/mappers/AllegationsOfHarmRevisedMapper.java
@@ -22,7 +22,7 @@ import javax.json.stream.JsonCollectors;
 
 import static java.util.Objects.nonNull;
 import static java.util.Optional.ofNullable;
-import static uk.gov.hmcts.reform.prl.mapper.citizen.CaseDataMapper.COMMA_SEPARATOR;
+import static uk.gov.hmcts.reform.prl.controllers.citizen.mapper.CaseDataMapper.COMMA_SEPARATOR;
 
 @Component
 public class AllegationsOfHarmRevisedMapper {
